+++ dz21/src/main/java/ru/otus/dz21/repository/BookRepository.java
@@ -14,14 +14,7 @@ import java.util.Optional;
 
 public interface BookRepository extends JpaRepository<Book, Integer> {
 
-    //    @PostFilter("hasPermission(filterObject.id, 'READ')")
-//    @PostFilter("hasPermission(filterObject, 'read') or hasPermission(filterObject, admin)")
-//    @PostFilter(
-//            value = "hasPermission(filterObject, 'READ')"
-////            filterTarget = "Book"
-//    )
-//    @PostAuthorize("hasPermission(returnObject, 'READ')")
-    @PostFilter("hasPermission(filterObject, 'READ')")
+//    @PostFilter("hasPermission(filterObject, 'READ')")
 //    @Secured({"ROLE_ADMIN", "ROLE_ADVANCED_USER"})
     List<Book> findAll();
 
@@ -34,8 +27,6 @@ public interface BookRepository extends JpaRepository<Book, Integer> {
 
     void deleteById(Integer id);
 
-    //    @PreAuthorize("hasPermission(#entityBook, 'WRITE')")
-//    @PreAuthorize("hasRole('ROLE_ADMIN') or hasPermission(#book, 'write')")
     @PreAuthorize("hasPermission(#book, 'write')")
     Book save(@Param("book") Book book);
