+++ dz21/src/test/java/ru/otus/dz21/controller/BookControllerTest.java
@@ -8,7 +8,12 @@ import org.junit.runner.RunWith;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.EmbeddedDatabaseConnection;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestConfiguration;
@@ -18,9 +23,16 @@ import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import ru.otus.dz21.domain.*;
 import ru.otus.dz21.dto.BookDto;
+import ru.otus.dz21.repository.AuthorRepository;
+import ru.otus.dz21.repository.BookRepository;
+import ru.otus.dz21.repository.GenreRepository;
 import ru.otus.dz21.repository.UserRepository;
 import ru.otus.dz21.security.AclAuthorizationConfiguration;
 import ru.otus.dz21.security.AclConfiguration;
@@ -39,19 +51,41 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(BookController.class)
-//@ConfigurationProperties("application")
-//@SpringBootTest
+//@WebMvcTest(BookController.class)
+@SpringBootTest
+//@ComponentScan("ru.otus.dz21.repository")
+//@SpringBootTest(classes = {SecurityConfiguration.class, AclConfiguration.class, AclAuthorizationConfiguration.class})//,//})
+//        DataSourceConf.class})
+@EnableWebMvc
+//@AutoConfigureTestDatabase(connection = EmbeddedDatabaseConnection.H2)
+//@AutoConfigureMockMvc(secure = false)
+//@Import({BookController.class, SecurityConfiguration.class, AclConfiguration.class, AclAuthorizationConfiguration.class})//,//})
+//        DataSourceConf.class})
+//@WebAppConfiguration
+//@SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
 @WithMockUser(
+//        username = "advuser",
         username = "user",
         authorities = {"ROLE_USER"}
+//        authorities = {"ROLE_ADVANCED_USER"}
 //        roles = {"USER"}
 //        authorities = {UserRoleEnum}
 )
 public class BookControllerTest {
 
     @Autowired
+    private WebApplicationContext wac;
+
+//    private MockMvc mockMvc;
     private MockMvc mvc;
+
+//    @Before
+//    public void setup() {
+//        this.mockMvc = MockMvcBuilders.webAppContextSetup(this.wac).build();
+//    }
+
+//    @Autowired
+//    private MockMvc mvc;
 
 //    @Autowired
 //    private UserDetailsService userDetailsService;
@@ -65,55 +99,64 @@ public class BookControllerTest {
     @MockBean
     private UserRepository userRepository;
 
+    @Autowired
+    private BookRepository bookRepositoryJpa;
+
+    @Autowired
+    private GenreRepository genreRepositoryJpa;
+
+    @Autowired
+    private AuthorRepository authorRepositoryJpa;
+
 //    @MockBean
 //    @Autowired
 //    private DataSource dataSource;
 
 
-    @Configuration
-    @ConfigurationProperties("application")
-    @ComponentScan(basePackageClasses = {BookController.class, SecurityConfiguration.class,
-            AclConfiguration.class, AclAuthorizationConfiguration.class})
-    public static class TestConf {
-
-        @Value("${spring.datasource.url}")
-        private String dataSourceUrl;
-
-        @Value("${spring.datasource.schema}")
-        private String dataSourceSchema;
-
-        //        private String url;
-        @Bean
-        DataSource dataSource() {
-            HikariConfig dataSourceConfig = new HikariConfig();
-//            dataSourceConfig.setJdbcUrl("jdbc:h2:mem:testdb");
-            dataSourceConfig.setJdbcUrl(dataSourceUrl);
-            dataSourceConfig.setSchema(dataSourceSchema);
-//            dataSourceConfig.set
-//            dataSourceConfig.setData
-//            dataSourceConfig.setUsername(env.getRequiredProperty("spring.datasource.username"));
-//            dataSourceConfig.setPassword(env.getRequiredProperty("spring.datasource.password"));
-            return new HikariDataSource(dataSourceConfig);
-        }
-
+//    @Configuration
+//    @ConfigurationProperties("application")
+//    @ComponentScan(basePackageClasses = {BookController.class, SecurityConfiguration.class,
+//            AclConfiguration.class, AclAuthorizationConfiguration.class})
+//    public static class TestConf {
+//
+//        @Value("${spring.datasource.url}")
+//        private String dataSourceUrl;
+//
+//        @Value("${spring.datasource.schema}")
+//        private String dataSourceSchema;
+//
+//        //        private String url;
 //        @Bean
-//        public DataSource dataSource() {
-//            return new HikariDataSource();
+//        DataSource dataSource() {
+//            HikariConfig dataSourceConfig = new HikariConfig();
+////            dataSourceConfig.setJdbcUrl("jdbc:h2:mem:testdb");
+//            dataSourceConfig.setJdbcUrl(dataSourceUrl);
+//            dataSourceConfig.setSchema(dataSourceSchema);
+////            dataSourceConfig.set
+////            dataSourceConfig.setData
+////            dataSourceConfig.setUsername(env.getRequiredProperty("spring.datasource.username"));
+////            dataSourceConfig.setPassword(env.getRequiredProperty("spring.datasource.password"));
+//            return new HikariDataSource(dataSourceConfig);
 //        }
-
-//        @Bean
-////        @Profile("test")
-//        public DataSource dataSource() {
-//            DriverManagerDataSource dataSource = new DriverManagerDataSource();
-//            dataSource.setDriverClassName("org.h2.Driver");
-//            dataSource.setUrl("jdbc:h2:mem:db;DB_CLOSE_DELAY=-1");
-//            dataSource.setUsername("sa");
-//            dataSource.setPassword("sa");
-//            dataSource.setSchema("data.sql");
-////            dataSource.("schema.sql");
-//            return dataSource;
-//        }
-    }
+//
+////        @Bean
+////        public DataSource dataSource() {
+////            return new HikariDataSource();
+////        }
+//
+////        @Bean
+//////        @Profile("test")
+////        public DataSource dataSource() {
+////            DriverManagerDataSource dataSource = new DriverManagerDataSource();
+////            dataSource.setDriverClassName("org.h2.Driver");
+////            dataSource.setUrl("jdbc:h2:mem:db;DB_CLOSE_DELAY=-1");
+////            dataSource.setUsername("sa");
+////            dataSource.setPassword("sa");
+////            dataSource.setSchema("data.sql");
+//////            dataSource.("schema.sql");
+////            return dataSource;
+////        }
+//    }
 
     private Author author;
     private Genre genre;
@@ -124,13 +167,21 @@ public class BookControllerTest {
 
     @Before
     public void setUp() throws Exception {
+
+        this.mvc = MockMvcBuilders.webAppContextSetup(this.wac).build();
+
         author = new Author("Лев", "Толстой");
         author.setId(1);
+        authorRepositoryJpa.save(author);
+
         genre = new Genre("роман-эпопея");
         genre.setId(1);
+        genreRepositoryJpa.save(genre);
+
         book = new Book("Война и мир", author, genre);
-        comment = new Comment("Эпично, но слишком затянуто.", book);
         book.setId(1);
+        bookRepositoryJpa.save(book);
+//        comment = new Comment("Эпично, но слишком затянуто.", book);
         books = Arrays.asList(book);
         bookDto = new BookDto(1, "Мертвые души", "Николай", "Гоголь", "поэма");
 
@@ -138,7 +189,7 @@ public class BookControllerTest {
 
     @Test
     public void booksPage() throws Exception {
-        Mockito.when(libraryService.listBooks()).thenReturn(books);
+//        Mockito.when(libraryService.listBooks()).thenReturn(books);
         mvc.perform(get("/books"))
                 .andExpect(status().isOk())
                 .andDo(print())
