+++ dz21/src/test/resources/application.yml
@@ -11,6 +11,7 @@ spring:
     driver-class-name: org.h2.Driver
     initialization-mode: always
     schema: classpath:schema.sql
+#    schema: schema.sql
 #    schema: classpath:/schema.sql
     data: classpath:data.sql
 #    data: data.sql
