+++ dz21/src/test/resources/data.sql
@@ -22,7 +22,7 @@ INSERT INTO acl_class (id, class) VALUES
 INSERT INTO acl_object_identity
   (id, object_id_class, object_id_identity, parent_object, owner_sid, entries_inheriting)
   VALUES
-  (1, 1, 1, NULL, 1, FALSE ),
+  (1, 1, 1, NULL, 1, FALSE );
 --   (2, 1, 2, NULL, 1, FALSE ),
 --   (3, 1, 3, NULL, 1, FALSE ),
 --   (4, 1, 4, NULL, 1, FALSE );
@@ -31,7 +31,8 @@ INSERT INTO acl_entry
   (id, acl_object_identity, ace_order, sid, mask, granting, audit_success, audit_failure)
   VALUES
   (1, 1, 1, 2, 0, TRUE , TRUE , TRUE ),
-  (2, 1, 2, 3, 1, TRUE , TRUE , TRUE );
+  (2, 1, 2, 3, 1, TRUE , TRUE , TRUE ),
+  (3, 1, 3, 3, 2, TRUE , TRUE , TRUE );
 --   (3, 2, 1, 2, 1, TRUE , TRUE , TRUE ),
 --   (4, 2, 2, 3, 1, TRUE , TRUE , TRUE ),
 --   (5, 3, 1, 2, 1, TRUE , TRUE , TRUE ),
